+++ controller/server/view/licensing_page.mli
@@ -1,0 +1,2 @@
+val html :
+  [> Html_types.html ] Tyxml.Html.elt Lwt.t