+++ controller/Changelog.md
@@ -4,6 +4,7 @@
 
 - kiosk: Add a key combination to perform hard refresh (Ctrl-Shift-R)
 - os: Added localization options for Polish and Czech
+- controller: Add licensing page in System Settings
 
 ## Changed